+++ clojure/walk/walk.go
@@ -82,9 +82,9 @@ func init() {
 							var v = cljs_core.Nth.X_invoke_Arity3(vec__4, float64(1), nil)
 							_, _, _ = vec__4, k, v
 							if reflect.ValueOf(k).Kind() == reflect.String {
-								return (&cljs_core.CljsCorePersistentVector{nil, 2, 5, cljs_core.CljsCorePersistentVector_EMPTY_NODE, []interface{}{cljs_core.Keyword.X_invoke_Arity1(k), v}, nil})
+								return (&cljs_core.CljsCorePersistentVector{nil, float64(2), float64(5), cljs_core.CljsCorePersistentVector_EMPTY_NODE, []interface{}{cljs_core.Keyword.X_invoke_Arity1(k), v}, nil})
 							} else {
-								return (&cljs_core.CljsCorePersistentVector{nil, 2, 5, cljs_core.CljsCorePersistentVector_EMPTY_NODE, []interface{}{k, v}, nil})
+								return (&cljs_core.CljsCorePersistentVector{nil, float64(2), float64(5), cljs_core.CljsCorePersistentVector_EMPTY_NODE, []interface{}{k, v}, nil})
 							}
 						}
 					})
@@ -121,9 +121,9 @@ func init() {
 							var v = cljs_core.Nth.X_invoke_Arity3(vec__10, float64(1), nil)
 							_, _, _ = vec__10, k, v
 							if func() bool { _, instanceof := k.(*cljs_core.CljsCoreKeyword); return instanceof }() {
-								return (&cljs_core.CljsCorePersistentVector{nil, 2, 5, cljs_core.CljsCorePersistentVector_EMPTY_NODE, []interface{}{cljs_core.Name.X_invoke_Arity1(k), v}, nil})
+								return (&cljs_core.CljsCorePersistentVector{nil, float64(2), float64(5), cljs_core.CljsCorePersistentVector_EMPTY_NODE, []interface{}{cljs_core.Name.X_invoke_Arity1(k), v}, nil})
 							} else {
-								return (&cljs_core.CljsCorePersistentVector{nil, 2, 5, cljs_core.CljsCorePersistentVector_EMPTY_NODE, []interface{}{k, v}, nil})
+								return (&cljs_core.CljsCorePersistentVector{nil, float64(2), float64(5), cljs_core.CljsCorePersistentVector_EMPTY_NODE, []interface{}{k, v}, nil})
 							}
 						}
 					})
